Prototype user auth support with supertokens
Used the NextJs + Email&Social login docs at:
https://supertokens.com/docs/thirdpartyemailpassword/nextjs/about
I haven't done step 5b to support auto-refresh of tokens because its not
obvious to me that its needed.

For development, set up the supertoken docker container and connected it
to the postgres docker container.

For production, we can use the same docker container (or we can use
supertoken-as-a-service).

- Added email verification using supertokens server for emails
- Comment out a social platforms till we are ready to set up our own
keys
- Add example code on checking if the user is logged in on the frontend